+++ contracts/interfaces/IMainPool.sol
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.2;
 
+// Author: Francesco Sullo <francesco@sullo.co>
+// (c) 2022+ SuperPower Labs Inc.
+
 interface IMainPool {
   event DepositSaved(address user, uint16 mainIndex);
 
@@ -17,7 +20,7 @@ interface IMainPool {
     // SYNR maxTokenSupply is 10 billion * 18 decimals = 1e28
     // which is less type(uint96).max (~79e28)
     uint96 tokenAmountOrID;
-    uint32 unlockedAt;
+    uint32 unstakedAt;
     uint16 otherChain;
     // since the process is asyncronous, the same deposit can be at a different mainIndex
     // on the main net and on the sidechain.
