+++ contracts/interfaces/ISidePool.sol
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.2;
 
+// Author: Francesco Sullo <francesco@sullo.co>
+// (c) 2022+ SuperPower Labs Inc.
+
 interface ISidePool {
   event DepositSaved(address user, uint16 mainIndex);
 
@@ -17,7 +20,7 @@ interface ISidePool {
     // SYNR maxTokenSupply is 10 billion * 18 decimals = 1e28
     // which is less type(uint96).max (~79e28)
     uint96 tokenAmountOrID;
-    uint32 unlockedAt;
+    uint32 unstakedAt;
     // @dev mainIndex Since the process is asyncronous, the same deposit can be at a different index
     // on the main net and on the sidechain. This guarantees alignment
     uint16 mainIndex;
@@ -31,7 +34,9 @@ interface ISidePool {
 
   /// @dev Data structure representing token holder using a pool
   struct User {
-    // @dev Total blueprints staked
+    // @dev Total passes staked for boost
+    uint16 passAmount;
+    // @dev Total blueprints staked for boost
     uint16 blueprintsAmount;
     // @dev Total staked amount
     uint256 tokenAmount;
@@ -47,6 +52,15 @@ interface ISidePool {
     uint32 lastRatioUpdateAt;
     uint16 swapFactor;
     uint16 stakeFactor;
+    uint16 taxPoints; // ex 250 = 2.5%
+  }
+
+  struct NftConf {
+    uint32 synrEquivalent; // 100,000
+    uint16 sPBoostFactor; // 1250 > 12.5%
+    uint32 sPBoostLimit;
+    uint16 bPBoostFactor;
+    uint32 bPBoostLimit;
   }
 
   function initPool(
@@ -54,7 +68,27 @@ interface ISidePool {
     uint32 decayInterval_,
     uint16 decayFactor_,
     uint16 swapFactor_,
-    uint16 stakeFactor_
+    uint16 stakeFactor_,
+    uint16 taxPoints_
+  ) external;
+
+  function updateConf(
+    uint32 decayInterval_,
+    uint16 decayFactor_,
+    uint16 swapFactor_,
+    uint16 stakeFactor_,
+    uint16 taxPoints_
+  ) external;
+
+  // Split configuration in two struct to avoid following error calling initPool
+  // CompilerError: Stack too deep when compiling inline assembly:
+  // Variable value0 is 1 slot(s) too deep inside the stack.
+  function updateNftConf(
+    uint32 synrEquivalent_,
+    uint16 sPBoostFactor_,
+    uint32 sPBoostLimit_,
+    uint16 bPBoostFactor_,
+    uint32 bPBoostLimit_
   ) external;
 
   function multiplier() external pure returns (uint256);
@@ -78,37 +112,9 @@ interface ISidePool {
 
   function getDepositIndexByMainIndex(address user, uint256 mainIndex) external view returns (uint256);
 
-  function withdrawPenalties(uint256 amount, address beneficiary) external;
-
-  // pool functions
-
-  //  function poolToken() external view returns (address);
-  //
-  //  function weight() external view returns (uint32);
-  //
-  //  function lastYieldDistribution() external view returns (uint64);
-  //
-  //  function yieldRewardsPerWeight() external view returns (uint256);
-  //
-  //  function usersLockingWeight() external view returns (uint256);
-  //
-  //  function pendingYieldRewards(address _user) external view returns (uint256);
-  //
-  //  function balanceOf(address _user) external view returns (uint256);
-  //
-  //  function stake(
-  //    uint256 amount,
-  //    uint64 lockedUntil
-  //  ) external;
-
-  //  function unstake(
-  //    uint256 depositIndex,
-  //    uint256 amount
-  //  ) external;
-
-  //  function sync() external;
-  //
-  //  function processRewards(bool useSSYN) external;
-  //
-  //  function setWeight(uint32 _weight) external;
+  function withdrawPenaltiesOrTaxes(
+    uint256 amount,
+    address beneficiary,
+    uint256 what
+  ) external;
 }
