+++ contracts/mocks/SeedFarmMock.sol
@@ -1,11 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.2;
 
-//import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
-//import "@openzeppelin/contracts-upgradeable/utils/AddressUpgradeable.sol";
-//import "@openzeppelin/contracts-upgradeable/utils/math/SafeMathUpgradeable.sol";
-//import "@ndujalabs/wormhole-tunnel/contracts/WormholeTunnelUpgradeable.sol";
-
 import "../SeedFarm.sol";
 import "hardhat/console.sol";
