+++ contracts/mocks/WormholeMock.sol
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.2;
 
+// Author: Francesco Sullo <francesco@sullo.co>
+// (c) 2022+ SuperPower Labs Inc.
+
 // the goal of this mock is to allow testing.
 // We assume that the bridging is successful
