--- test/Payload.test.js
@@ -1,322 +1,0 @@
-const {expect, assert, use} = require("chai");
-
-const {initEthers, assertThrowsMessage, getTimestamp, increaseBlockTimestampBy, bytes32Address, BNMulBy} = require("./helpers");
-const {upgrades} = require("hardhat");
-
-// tests to be fixed
-
-function normalize(val, n = 18) {
-  return "" + val + "0".repeat(n);
-}
-
-// test unit coming soon
-
-describe("#Payload", function () {
-  let WormholeMock, wormhole;
-  let SyndicateERC20, synr;
-  let SyntheticSyndicateERC20, sSynr;
-  let SynrBridge, synrBridge;
-  let SynrBridgeV2;
-  let SeedFarm, seedFarm;
-  let SideToken, seed;
-  let SynCityPasses, pass;
-
-  const BN = ethers.BigNumber.from;
-
-  let deployer, fundOwner, superAdmin, operator, validator, user1, user2, marketplace, treasury;
-
-  before(async function () {
-    initEthers(ethers);
-    [deployer, fundOwner, superAdmin, operator, validator, user1, user2, marketplace, treasury] = await ethers.getSigners();
-    SyndicateERC20 = await ethers.getContractFactory("SyndicateERC20");
-    SyntheticSyndicateERC20 = await ethers.getContractFactory("SyntheticSyndicateERC20");
-    SynrBridge = await ethers.getContractFactory("SynrBridgeMock");
-    SynrBridgeV2 = await ethers.getContractFactory("SynrBridgeV2Mock");
-    SeedFarm = await ethers.getContractFactory("SeedFarmMock");
-    SideToken = await ethers.getContractFactory("SideToken");
-    WormholeMock = await ethers.getContractFactory("WormholeMock");
-    SynCityPasses = await ethers.getContractFactory("SynCityPasses");
-  });
-
-  async function initAndDeploy() {
-    const maxTotalSupply = 10000000000; // 10 billions
-    synr = await SyndicateERC20.deploy(fundOwner.address, maxTotalSupply, superAdmin.address);
-    await synr.deployed();
-    let features =
-      (await synr.FEATURE_TRANSFERS_ON_BEHALF()) +
-      (await synr.FEATURE_TRANSFERS()) +
-      (await synr.FEATURE_UNSAFE_TRANSFERS()) +
-      (await synr.FEATURE_DELEGATIONS()) +
-      (await synr.FEATURE_DELEGATIONS_ON_BEHALF());
-    await synr.updateFeatures(features);
-
-    sSynr = await SyntheticSyndicateERC20.deploy(superAdmin.address);
-    await sSynr.deployed();
-
-    pass = await SynCityPasses.deploy(validator.address);
-    await pass.deployed();
-
-    synrBridge = await upgrades.deployProxy(SynrBridge, [synr.address, sSynr.address, pass.address]);
-    await synrBridge.deployed();
-
-    await sSynr.updateRole(synrBridge.address, await sSynr.ROLE_WHITE_LISTED_RECEIVER());
-
-    seed = await upgrades.deployProxy(SideToken, ["Mobland SEED Token", "SEED"]);
-    await seed.deployed();
-
-    seedFarm = await upgrades.deployProxy(SeedFarm, [seed.address]);
-    await seedFarm.deployed();
-
-    await seed.grantRole(await seed.MINTER_ROLE(), seedFarm.address);
-
-    wormhole = await WormholeMock.deploy();
-    await synrBridge.wormholeInit(2, wormhole.address);
-    await wormhole.deployed();
-
-    await synrBridge.wormholeRegisterContract(4, bytes32Address(seedFarm.address));
-    await synrBridge.initPool(7, 4000);
-
-    await seedFarm.wormholeInit(4, wormhole.address);
-    await seedFarm.wormholeRegisterContract(2, bytes32Address(synrBridge.address));
-  }
-
-  describe("#serializeInput", async function () {
-    beforeEach(async function () {
-      await initAndDeploy();
-    });
-
-    it("should serialize input", async function () {
-      const amount = ethers.utils.parseEther("10000");
-
-      const payload = await synrBridge.serializeInput(
-        1, // SYNR
-        365, // 1 year
-        amount
-      );
-
-      expect(payload).equal("100000000000000000000003651");
-    });
-
-    it("should throw invalid token", async function () {
-      const amount = ethers.utils.parseEther("10000");
-
-      expect(synrBridge.serializeInput(4, 365, amount)).revertedWith("Payload: invalid token type");
-    });
-
-    it("should throw not a mobland pass", async function () {
-      const amount = ethers.utils.parseEther("10000");
-
-      expect(synrBridge.serializeInput(2, 365, amount)).revertedWith("Payload: Not a Mobland SYNR Pass token ID");
-    });
-
-    it("should throw amount of range", async function () {
-      const amount = ethers.utils.parseEther("1000000000000");
-
-      expect(synrBridge.serializeInput(1, 365, amount)).revertedWith("Payload: tokenAmountOrID out of range");
-    });
-
-    it("should throw lockedTime out of range", async function () {
-      const amount = ethers.utils.parseEther("10000");
-
-      expect(synrBridge.serializeInput(1, 1e5, amount)).revertedWith("Payload: lockedTime out of range");
-    });
-  });
-
-  describe("#deserializeInput", async function () {
-    beforeEach(async function () {
-      await initAndDeploy();
-    });
-
-    it("should deserialize", async function () {
-      const amount = ethers.utils.parseEther("10000");
-
-      const payload = await synrBridge.serializeInput(
-        1, // SYNR
-        365, // 1 year
-        amount
-      );
-      const deserialize = await synrBridge.deserializeInput(payload);
-
-      expect(parseInt(deserialize)).equal(1, 365, amount);
-    });
-
-    // TODO add a fake payload and verify if it fails
-  });
-
-  describe("#Deposit", async function () {
-    beforeEach(async function () {
-      await initAndDeploy();
-    });
-
-    it("should return length of deposits", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      await synr.connect(fundOwner).transferFrom(fundOwner.address, user1.address, amount);
-      const payload = "100000000000000000000003651"
-      await synr.connect(user1).approve(synrBridge.address, ethers.utils.parseEther("10000"));
-      expect(
-        await synrBridge.connect(user1).wormholeTransfer(
-          payload,
-          4, // BSC
-          bytes32Address(user1.address),
-          1
-        )
-      )
-        .emit(synrBridge, "DepositSaved")
-        .withArgs(user1.address, 0);
-      const lenght = await synrBridge.getDepositsLength(user1.address);
-      expect(parseInt(lenght)).equal(1);
-    });
-
-    it("should return deposit by index", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      await synr.connect(fundOwner).transferFrom(fundOwner.address, user1.address, amount);
-      const payload = "100000000000000000000003651"
-      await synr.connect(user1).approve(synrBridge.address, ethers.utils.parseEther("10000"));
-      expect(
-        await synrBridge.connect(user1).wormholeTransfer(
-          payload,
-          4, // BSC
-          bytes32Address(user1.address),
-          1
-        )
-      )
-        .emit(synrBridge, "DepositSaved")
-        .withArgs(user1.address, 0);
-      const deposit = await synrBridge.getDepositByIndex(user1.address, 0);
-      expect(parseInt(deposit)).equal(1, deposit.lockedFrom, deposit.lockedUntil, 0, amount);
-    });
-  });
-
-  describe("#fromDepositToTransferPayload", async function () {
-    it("should from deposit to transfer payload", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      const lockedFrom = await getTimestamp();
-      const lockedUntil = lockedFrom + 3600 * 24 * 180;
-      const deposit = {
-        tokenType: 1,
-        lockedFrom,
-        lockedUntil,
-        tokenAmountOrID: amount,
-        unlockedAt: 0,
-        otherChain: 4,
-        mainIndex: 0,
-      };
-
-      const expected = BN(1)
-          .add(await BNMulBy(lockedFrom, 10))
-          .add(await BNMulBy(lockedUntil, 1, 11))
-          .add(await BNMulBy(0, 1, 21))
-          .add(await BNMulBy(amount, 1, 26));
-      const payload = await synrBridge.fromDepositToTransferPayload(deposit);
-      expect(payload).equal(expected)
-    });
-
-    it("should throw for invalid token type", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      const lockedFrom = await getTimestamp();
-      const lockedUntil = lockedFrom + 3600 * 24 * 180;
-      const deposit = {
-        tokenType: 3,
-        lockedFrom,
-        lockedUntil,
-        tokenAmountOrID: amount,
-        unlockedAt: 0,
-        otherChain: 4,
-        mainIndex: 0,
-      };
-      expect(synrBridge.fromDepositToTransferPayload(deposit)).revertedWith("Payload: invalid token type")
-    });
-
-    it("should throw invalid interval", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      const lockedFrom = await getTimestamp();
-      const lockedUntil = 1;
-      const deposit = {
-        tokenType: 1,
-        lockedFrom,
-        lockedUntil,
-        tokenAmountOrID: amount,
-        unlockedAt: 0,
-        otherChain: 4,
-        mainIndex: 0,
-      };
-      expect(synrBridge.fromDepositToTransferPayload(deposit)).revertedWith("Payload: invalid interval")
-    });
-
-    it("should throw tokenAmount out of range", async function () {
-      const amount = ethers.utils.parseEther("10000000000");
-      const lockedFrom = await getTimestamp();
-      const lockedUntil = lockedFrom + 3600 * 24 * 180;
-      const deposit = {
-        tokenType: 1,
-        lockedFrom,
-        lockedUntil,
-        tokenAmountOrID: amount,
-        unlockedAt: 0,
-        otherChain: 4,
-        mainIndex: 0,
-      };
-      expect(synrBridge.fromDepositToTransferPayload(deposit)).revertedWith("Payload: tokenAmountOrID out of range")
-    });
-  });
-
-  describe("#deserializeDeposit", async function () {
-    beforeEach(async function () {
-      await initAndDeploy();
-    });
-
-    it("should deserialize deposit", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      await synr.connect(fundOwner).transferFrom(fundOwner.address, user1.address, amount);
-      const payload = "100000000000000000000003651"
-      await synr.connect(user1).approve(synrBridge.address, ethers.utils.parseEther("10000"));
-      expect(
-        await synrBridge.connect(user1).wormholeTransfer(
-          payload,
-          4, // BSC
-          bytes32Address(user1.address),
-          1
-        )
-      )
-        .emit(synrBridge, "DepositSaved")
-        .withArgs(user1.address, 0);
-      const deposit = await synrBridge.getDepositByIndex(user1.address, 0);
-      const deserialize = await synrBridge.deserializeDeposit(parseInt(deposit));
-      expect(parseInt(deserialize)).equal(1, deposit.lockedFrom, deposit.lockedUntil, 0, amount);
-    });
-  });
-
-  describe("#_updateUserAndAddDeposit", async function () {
-    beforeEach(async function () {
-      await initAndDeploy();
-    });
-
-    it("should return updated user", async function () {
-      const amount = ethers.utils.parseEther("10000");
-      await synr.connect(fundOwner).transferFrom(fundOwner.address, user1.address, amount);
-      const payload = "100000000000000000000003651"
-      await synr.connect(user1).approve(synrBridge.address, ethers.utils.parseEther("10000"));
-      expect(
-        await synrBridge.connect(user1).wormholeTransfer(
-          payload,
-          4, // BSC
-          bytes32Address(user1.address),
-          1
-        )
-      )
-        .emit(synrBridge, "DepositSaved")
-        .withArgs(user1.address, 0);
-      await increaseBlockTimestampBy(182.5 * 24 * 3600);
-      await synrBridge.updateUserAndAddDeposit(user1.address, 1, 1000000000, 3000000000, amount, 44, 0);
-      //Update user pushes new deposit, it therefore changes the index of the intended new update deposite to the last one in the list.
-      //unsure if that is the intended behavior of UPDATE USER
-      const depositAfter = await synrBridge.getDepositByIndex(user1.address, 1);
-      expect(depositAfter.tokenType, depositAfter.lockedFrom, depositAfter.lockedUntil, depositAfter.otherChain).equal(
-        1,
-        1000000000,
-        3000000000,
-        44
-      );
-    });
-  });
-});
